Add languagetool check via reviewdog (#8)
https://github.com/reviewdog/action-languagetool

https://languagetool.org/

https://github.com/languagetool-org/languagetool

Use `en` locale as it does not do spell checking, whereas if we use
`en-gb`then it flags errors for the US spelling of `LICENSE`
(which is what it should do).
As LICENSE is the conventional name for the licence file, we
don't want that to be reported as an error.

Our spell checking is done by misspell (which we have set to allow UK
or US variants e.g. colour and color), so we don't want languagetool
to do spell checking anyway.  It will still do grammar checks which is
what we want.